Organizing Typescipt API (#2287)
First I wanna keep every helper using typescript factory in one place.
This should allow to use "integration helpers" file for bigger and
higher order entities of the code production.
That will be helpful for #2280 that makes "integraton" file even bigger
and less readable.
