+++ eslint.config.js
@@ -155,7 +155,7 @@ export default tsPlugin.config(
   },
   {
     name: "source/integration",
-    files: ["src/integration.ts"],
+    files: ["src/integration.ts", "src/zts.ts"],
     rules: {
       "no-restricted-syntax": [
         "warn",
