+++ src/zts.ts
@@ -11,11 +11,9 @@ import { HandlingRules, walkSchema } from "./schema-walker";
 import {
   addJsDocComment,
   isPrimitive,
-  LiteralType,
-  makePropertyIdentifier,
-  Producer,
-  ZTSContext,
-} from "./zts-helpers";
+  makeInterfaceProp,
+} from "./typescript-api";
+import { LiteralType, Producer, ZTSContext } from "./zts-helpers";
 
 const { factory: f } = ts;
 
@@ -53,14 +51,9 @@ const onObject: Producer = (
       isResponse && hasCoercion(value)
         ? value instanceof z.ZodOptional
         : value.isOptional();
-    const propertySignature = f.createPropertySignature(
-      undefined,
-      makePropertyIdentifier(key),
-      isOptional && hasQuestionMark
-        ? f.createToken(ts.SyntaxKind.QuestionToken)
-        : undefined,
-      next(value),
-    );
+    const propertySignature = makeInterfaceProp(key, next(value), {
+      isOptional: isOptional && hasQuestionMark,
+    });
     return value.description
       ? addJsDocComment(propertySignature, value.description)
       : propertySignature;
